1.5.2: Some fixes

+++ Rogue Re-Rezzed/src/ui/Status.java
@@ -30,7 +30,7 @@ public class Status extends JPanel{
         g2.fillRect(10, 65, 100*Player.xp/(10*Player.xplevels), 20);
         g2.setColor(Color.WHITE);
         g2.drawString("You are on level: "+Level.numLevels, 10, 20);
-        g2.drawString("Health: "+Rogue.getLevel().getPlayer().health, 10, 40);
+        g2.drawString("Health: "+(int) Rogue.getLevel().getPlayer().health, 10, 40);
         g2.drawString("Mana: "+Rogue.getLevel().getPlayer().mana, 10, 60);
         g2.drawString("You are Level: "+Player.xplevels, 10, 80);
         g2.drawString("Max Attack: "+Rogue.getLevel().getPlayer().maxAtt, 10, 100);
